refactor: 중복 로직 hook으로 분리 및 오타 수정
