+++ src/components/my-page/MyPageEditProfileImage.tsx
@@ -1,46 +1,21 @@
-import { useRef } from 'react';
 import PlusCircle from '@/assets/plus-circle.svg';
 import { ThumbnailUploadProps } from '@/types/thumbnail';
+import { useSelectImage } from '@/hooks/useSelectImage';
 
 const MyPageEditProfileImage = ({
   userImage,
   onImageChange,
   imgFile,
 }: ThumbnailUploadProps & { userImage: string }) => {
-  const imgRef = useRef<HTMLInputElement>(null);
-
-  const saveImgFile = () => {
-    if (imgRef.current?.files && imgRef.current.files[0]) {
-      const file = imgRef.current.files[0];
-      const reader = new FileReader();
-
-      reader.readAsDataURL(file);
-      reader.onloadend = () => {
-        if (reader.result) {
-          onImageChange(reader.result as string);
-        }
-      };
-    }
-  };
-  const handleInput = () => {
-    imgRef.current?.click();
-  };
+  const { imgRef, saveImgFile, handleInput } = useSelectImage(onImageChange);
   return (
     <section className="flex flex-col items-center">
       <div className="h-24 w-24 overflow-hidden rounded-full">
-        {imgFile ? (
-          <img
-            src={imgFile}
-            alt="프로필 사진"
-            className="w-ull h-full object-cover"
-          />
-        ) : (
-          <img
-            src={userImage}
-            alt="프로필 사진"
-            className="w-ull h-full object-cover"
-          />
-        )}
+        <img
+          src={imgFile ? imgFile : userImage}
+          alt="프로필 사진"
+          className="h-full w-full object-cover"
+        />
       </div>
 
       <div className="relative bottom-10 left-10 cursor-pointer overflow-hidden rounded-full bg-white">
