+++ src/pages/VideoEditPage.tsx
@@ -69,7 +69,7 @@ const VideoEditPage = () => {
           className="w-1/2"
           onClick={handleImageReset}
         >
-          취소
+          초기화
         </Button>
       </div>
     </main>
