+++ src/pages/my-page/MyPageEdit.tsx
@@ -42,7 +42,7 @@ const MyPageEdit = () => {
           className="w-1/2"
           onClick={handleImageReset}
         >
-          취소
+          초기화
         </Button>
       </div>
     </figure>
